Accomodate contract changes, make registry entry point for subgraph
