+++ subgraph/abis/EagletFactory.json
@@ -1,1 +1,1 @@
-[{"inputs":[{"internalType":"contract OptimisticQueueFactory","name":"_queueFactory","type":"address"}],"stateMutability":"nonpayable","type":"constructor"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"contract Eaglet","name":"eaglet","type":"address"},{"indexed":false,"internalType":"contract OptimisticQueue","name":"queue","type":"address"}],"name":"NewEaglet","type":"event"},{"inputs":[],"name":"newDummyEaglet","outputs":[{"internalType":"contract Eaglet","name":"eaglet","type":"address"},{"internalType":"contract OptimisticQueue","name":"queue","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"queueFactory","outputs":[{"internalType":"contract OptimisticQueueFactory","name":"","type":"address"}],"stateMutability":"view","type":"function"}]
+[{"inputs":[{"internalType":"contract ERC3000Registry","name":"_registry","type":"address"},{"internalType":"contract OptimisticQueueFactory","name":"_queueFactory","type":"address"}],"stateMutability":"nonpayable","type":"constructor"},{"inputs":[{"internalType":"string","name":"_name","type":"string"}],"name":"newDummyEaglet","outputs":[{"internalType":"contract Eaglet","name":"eaglet","type":"address"},{"internalType":"contract OptimisticQueue","name":"queue","type":"address"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"queueFactory","outputs":[{"internalType":"contract OptimisticQueueFactory","name":"","type":"address"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"registry","outputs":[{"internalType":"contract ERC3000Registry","name":"","type":"address"}],"stateMutability":"view","type":"function"}]
