+++ subgraph/scripts/build-manifest.sh
@@ -15,7 +15,7 @@ echo 'Generating manifest from data file: '$DATA
 cat $DATA
 
 mustache \
-  -p manifest/templates/sources/EagletFactory.yaml \
-  -p manifest/templates/contracts/EagletFactory.template.yaml \
+  -p manifest/templates/sources/ERC3000Registry.yaml \
+  -p manifest/templates/contracts/ERC3000Registry.template.yaml \
   $DATA \
   subgraph.template.yaml > subgraph.yaml
