+++ subgraph/src/OptimisticQueue.ts
@@ -1,7 +1,6 @@
 import { Configured as ConfiguredEvent } from '../generated/templates/OptimisticQueue/OptimisticQueue'
 import { OptimisticQueue, Config, Collateral } from '../generated/schema'
 
-
 export function handleConfigured(event: ConfiguredEvent): void {
   let queue = OptimisticQueue.load(event.address.toHexString())
   // TODO: Can there be no queue? check event processing order
